pyproject.toml
[tool.poetry]
name = "crmservice"
version = "1.0.0"
description = "CRMService for DevRel Hack 2.0"
authors = ["It Soda <nikshchegilskiy@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
djoser = "^2.2.2"
pillow = "^10.1.0"
django-debug-toolbar = "^4.2.0"
django-environ = "0.10.0"
django-extensions = "3.2.3"
django-redis = "5.3.0"
psycopg2-binary = "2.9.3"
redis = "4.6.0"
pytelegrambotapi = "4.13.0"
drf-extensions = "0.7.1"
mysqlclient = "2.1.1"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.1"
environ = "^1.0"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
flake8 = "5.0.4"
isort = "5.10.1"
black = "23.10.1"
coverage = "7.3.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"